refactor: rename `fetchActions.ts` to `apiRequests.ts`
