+++ src/entityOperations/createEntitiesForTool.ts
@@ -2,7 +2,7 @@ import * as R from "ramda";
 import type { SagaIterator } from "redux-saga";
 import { call, delay, put, select } from "redux-saga/effects";
 
-import { getApiDelayForTool } from "~/entityOperations/fetchActions";
+import { getApiDelayForTool } from "~/entityOperations/apiRequests";
 import { incrementEntityGroupTransferCompletedCount } from "~/modules/allEntities/allEntitiesActions";
 import { workspaceIdToLinkedIdSelector } from "~/modules/workspaces/workspacesSelectors";
 import { EntityGroup, ToolName, ValidEntity } from "~/typeDefs";
