+++ src/entityOperations/fetchEntitiesForTool.ts
@@ -2,7 +2,7 @@ import * as R from "ramda";
 import type { SagaIterator } from "redux-saga";
 import { call, delay, select } from "redux-saga/effects";
 
-import { getApiDelayForTool } from "~/entityOperations/fetchActions";
+import { getApiDelayForTool } from "~/entityOperations/apiRequests";
 import { mappingByToolNameSelector } from "~/modules/allEntities/allEntitiesSelectors";
 import { includedWorkspaceIdsByMappingSelector } from "~/modules/workspaces/workspacesSelectors";
 import { ToolName } from "~/typeDefs";
