+++ src/modules/clients/sagas/togglClientsSagas.ts
@@ -1,13 +1,13 @@
 import type { SagaIterator } from "redux-saga";
 import { call } from "redux-saga/effects";
 
-import { createEntitiesForTool } from "~/entityOperations/createEntitiesForTool";
-import { deleteEntitiesForTool } from "~/entityOperations/deleteEntitiesForTool";
 import {
   fetchArray,
   fetchEmpty,
   fetchObject,
-} from "~/entityOperations/fetchActions";
+} from "~/entityOperations/apiRequests";
+import { createEntitiesForTool } from "~/entityOperations/createEntitiesForTool";
+import { deleteEntitiesForTool } from "~/entityOperations/deleteEntitiesForTool";
 import { fetchEntitiesForTool } from "~/entityOperations/fetchEntitiesForTool";
 import { ClientModel, EntityGroup, ToolName } from "~/typeDefs";
 import { validStringify } from "~/utilities/textTransforms";
