+++ src/modules/users/sagas/togglUsersSagas.ts
@@ -6,7 +6,7 @@ import {
   fetchEmpty,
   fetchObject,
   getApiDelayForTool,
-} from "~/entityOperations/fetchActions";
+} from "~/entityOperations/apiRequests";
 import { fetchEntitiesForTool } from "~/entityOperations/fetchEntitiesForTool";
 import { incrementEntityGroupTransferCompletedCount } from "~/modules/allEntities/allEntitiesActions";
 import { includedSourceProjectIdsSelector } from "~/modules/projects/projectsSelectors";
