+++ src/modules/workspaces/sagas/clockifyWorkspacesSagas.ts
@@ -1,7 +1,7 @@
 import type { SagaIterator } from "redux-saga";
 import { call } from "redux-saga/effects";
 
-import { fetchArray } from "~/entityOperations/fetchActions";
+import { fetchArray } from "~/entityOperations/apiRequests";
 import { ClockifyMembershipResponseModel } from "~/modules/users/sagas/clockifyUsersSagas";
 import { EntityGroup, WorkspaceModel } from "~/typeDefs";
