+++ src/modules/workspaces/sagas/togglWorkspacesSagas.ts
@@ -1,7 +1,7 @@
 import type { SagaIterator } from "redux-saga";
 import { call } from "redux-saga/effects";
 
-import { fetchArray } from "~/entityOperations/fetchActions";
+import { fetchArray } from "~/entityOperations/apiRequests";
 import { EntityGroup, WorkspaceModel } from "~/typeDefs";
 
 export interface TogglWorkspaceResponseModel {
